install/requirements.txt
ipykernel>=6.29.0
jupyterlab>=4.1.0
notebook>=7.1.0
requests>=2.31.0
numpy>=1.26.0
pandas>=2.2.0
python-dotenv>=1.0.1
elasticsearch>=8.12.0
